Single element hovering
Organised src folder.
Added buffer methods for OnEvents which shouldn't be called due to properties.
e.g. OnSizeChange shouldn't be called if width or height is <= 0.
Added to TODO.md
Hover and selection still has a bug with elements that are invisable,
This will be fixed when I add depth management.